+++ src/collective/weather/tests/test_doctest.py
@@ -18,4 +18,12 @@ def test_suite():
         layered(doctest.DocTestSuite('collective.weather.utilities.yahoo'),
                 layer=FUNCTIONAL_TESTING),
     ])
+    suite.addTests([
+        layered(doctest.DocTestSuite('collective.weather.utilities.forecastio'),
+                layer=FUNCTIONAL_TESTING),
+    ])
+    suite.addTests([
+        layered(doctest.DocTestSuite('collective.weather.utilities.wunderground'),
+                layer=FUNCTIONAL_TESTING),
+    ])
     return suite